[HAPI-177] Implementing new class as index
